requirements.txt
Django==1.7.7
psycopg2==2.6
django-nose==1.3
selenium==2.44.0
factory-boy==2.4.1
coverage==3.7.1

